requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.3
Flask-WTF==1.2.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
email-validator==2.2.0
gunicorn==21.2.0
Flask-Mail==0.9.1
Werkzeug==2.3.0
urllib3==1.26.15
requests==2.31.0
# GCP specific
google-cloud-storage==2.13.0
google-cloud-logging==3.8.0
google-cloud-secret-manager==2.16.4
